+++ tabbycat/results/mixins.py
@@ -1,6 +1,13 @@
+from smtplib import SMTPException
+
+from django.contrib import messages
+from django.utils.translation import gettext as _
+
 from utils.misc import get_ip_address
 
 from .models import Submission
+from .result import DebateResult
+from .utils import send_ballot_receipt_emails_to_adjudicators
 
 
 class TabroomSubmissionFieldsMixin:
@@ -26,3 +33,17 @@ class PublicSubmissionFieldsMixin:
             'submitter_type': Submission.SUBMITTER_PUBLIC,
             'ip_address': get_ip_address(self.request)
         }
+
+
+class BallotEmailWithStatusMixin:
+    def send_email_receipts(self):
+        try:
+            send_ballot_receipt_emails_to_adjudicators(DebateResult(self.ballotsub).as_dicts(), self.debate)
+        except SMTPException:
+            messages.error(self.request, _("There was a problem sending ballot receipts to adjudicators."))
+            return False
+        except ConnectionError:
+            messages.error(self.request, _("There was a problem connecting to the e-mail server when trying to send ballot receipts to adjudicators."))
+            return False
+        else:
+            return True
